requirements.txt
asgiref==3.5.2
Django==4.1.1
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-js-asset==2.0.0
djangorestframework==3.13.1
Pillow==9.2.0
psycopg2-binary==2.9.3
python-decouple==3.6
pytils==0.4.1
pytz==2022.2.1
sqlparse==0.4.2
